requirements.txt
discord.py
aiohttp
asyncio
python-dotenv
mistralai
requests
discord
matplotlib
openai
newspaper3k
lxml_html_clean